+++ ChangeLog.md
@@ -5,6 +5,7 @@
 - hide errors from sudo -h when using old sudo. [#88][]
 - don't detect sudo if running as root. fixes [#101][]
 - hpacucli: check for array status. fixes [#98][]
+- hp_msa: allow configuring via plugin-options. [#97][]
 
 ## 3.2.3 (2015-03-25)
 
@@ -167,3 +168,4 @@ NOTES:
 [#100]: https://github.com/glensc/nagios-plugin-check_raid/issues/100
 [#101]: https://github.com/glensc/nagios-plugin-check_raid/issues/101
 [#98]: https://github.com/glensc/nagios-plugin-check_raid/issues/98
+[#97]: https://github.com/glensc/nagios-plugin-check_raid/issues/97